Support legacy sanitizers options
