--- x-pack/filebeat/input/azureeventhub/sanitization.go
@@ -1,102 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License;
-// you may not use this file except in compliance with the Elastic License.
-
-//go:build !aix
-// +build !aix
-
-package azureeventhub
-
-import (
-	"bytes"
-	"errors"
-	"fmt"
-	"regexp"
-	"strings"
-)
-
-type sanitizationOption string
-
-const (
-	newLines     sanitizationOption = "NEW_LINES"
-	singleQuotes sanitizationOption = "SINGLE_QUOTES"
-	regExp       sanitizationOption = "REGEXP"
-)
-
-// sanitizeOptionsValidate validates for supported sanitization options
-func sanitizeOptionsValidate(s string) error {
-	switch s {
-	case "NEW_LINES":
-		return nil
-	case "SINGLE_QUOTES":
-		return nil
-	case "REGEXP":
-		return nil
-	default:
-		return errors.New("invalid sanitization option")
-	}
-}
-
-// sanitize applies the sanitization options specified in the config
-// if no sanitization options are provided, the message remains unchanged
-func sanitize(jsonByte []byte, opts ...string) []byte {
-	res := jsonByte
-
-	for _, opt := range opts {
-		switch sanitizationOption(opt) {
-		case newLines:
-			res = sanitizeNewLines(res)
-		case singleQuotes:
-			res = sanitizeSingleQuotes(res)
-		case regExp:
-			res = sanitizeRegex(res)
-		}
-	}
-
-	return res
-}
-
-// sanitizeNewLines removes newlines found in the message
-func sanitizeNewLines(jsonByte []byte) []byte {
-	return bytes.ReplaceAll(jsonByte, []byte("\n"), []byte{})
-}
-
-// sanitizeSingleQuotes replaces single quotes with double quotes in the message
-// single quotes that are in between double quotes remain unchanged
-func sanitizeSingleQuotes(jsonByte []byte) []byte {
-	var result bytes.Buffer
-	var prevChar byte
-
-	inDoubleQuotes := false
-
-	for _, r := range jsonByte {
-		if r == '"' && prevChar != '\\' {
-			inDoubleQuotes = !inDoubleQuotes
-		}
-
-		if r == '\'' && !inDoubleQuotes {
-			result.WriteRune('"')
-		} else {
-			result.WriteByte(r)
-		}
-		prevChar = r
-	}
-
-	return result.Bytes()
-}
-
-func sanitizeRegex(jsonByte []byte) []byte {
-	// Remove any leading/trailing whitespace
-	input := strings.TrimSpace(string(jsonByte))
-
-	// Regular expression to match array contents that are not valid JSON
-	// re := regexp.MustCompile(`\[\s*([^[\]{},\s]+(?:\s+[^[\]{},\s]+)*)\s*\]`)
-	re := regexp.MustCompile(`\[\s*([^\[\]{},\s]+(?:\s+[^\[\]{},\s]+)*)\s*\]`)
-
-	// Replace invalid array contents with a string placeholder
-	sanitized := re.ReplaceAllStringFunc(input, func(match string) string {
-		return fmt.Sprintf("[\"%s\"]", strings.TrimSpace(match[1:len(match)-1]))
-	})
-
-	return []byte(sanitized)
-}
